requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==2.0.4
behave==1.2.5
behave-django==0.4.0
billiard==3.3.0.23
biopython==1.67
blessings==1.6
celery==3.1.24
coverage==4.1
cssselect==0.9.2
dj-database-url==0.4.1
Django==1.9.7
django-celery==3.1.17
django-debug-toolbar==1.4
django-nose==1.4.4
django-randomslugfield==0.3.0
django-with-asserts==0.0.1
extras==1.0.0
fixtures==3.0.0
flake8==2.6.2
future==0.15.2
fuzzywuzzy==0.12.0
gherkin-official==4.0.0
gunicorn==19.6.0
isort==4.3.4
Jinja2==2.8
kombu==3.0.36
lazy-object-proxy==1.3.1
linecache2==1.0.0
lxml==3.6.0
MarkupSafe==0.23
mccabe==0.5.0
mock==2.0.0
nose==1.3.7
nose-exclude==0.4.1
numpy==1.11.1
parse==1.6.6
parse-type==0.3.4
pbr==1.10.0
pep8==1.7.0
psycopg2==2.7.5
pycodestyle==2.0.0
pyflakes==1.2.3
pylint==2.1.1
pyperclip==1.5.27
python-dateutil==2.5.3
python-mimeparse==1.5.5
python-subunit==1.2.0
pytz==2016.6.1
PyYAML==3.12
redis==2.10.5
repoze.lru==0.6
requests==2.10.0
selenium==2.53.6
six==1.10.0
sqlparse==0.1.19
sure==1.4.0
testtools==2.2.0
traceback2==1.4.0
typed-ast==1.1.0
unittest2==1.1.0
whitenoise==3.2.1
wrapt==1.10.11
